JGraph.jl

JGraph.Graph

JGraph.add_edge!

JGraph.add_vertex!

JGraph.degree

JGraph.fadj

JGraph.has_edge

JGraph.neighbors

JGraph.vertices

JGraph.Graph — Method

Graph(n, m)

generate a graph with n vertices and m edges. Return a graph G.

JGraph.add_edge! — Method

add_edge(G::AbstractGraph, u::T, v::T)

Add the egde(u,v) in the graph G. Return False is the edge already exit, true otherwise.

JGraph.add_vertex! — Method

add_edge(G::AbstractGraph, num_vertex)

Add "num_vertex" vertices to the graph G.

JGraph.degree — Method

degree(G::Graph)

Return the list of nodes degree.

JGraph.has_edge — Method

has_edge(G::AbstractGraph, u::T, v::T)

Test if egde(u,v) exist in the graph G.

JGraph.neighbors — Method

neighbors(G::Graph, u)

return the neighbors list adjacent to the node u

JGraph.vertices — Method

vertices(G::Graph)

vertices itertor

JGraph.fadj — Method

fadj(G::Graph)

return the adjacency list of the graph